+++ lib/internal/bootstrap/pre_execution.js
@@ -36,9 +36,6 @@ function prepareMainThreadExecution(expandArgv1 = false) {
 
   setupDebugEnv();
 
-  // Only main thread receives signals.
-  setupSignalHandlers();
-
   // Process initial diagnostic reporting configuration, if present.
   initializeReport();
   initializeReportSignalHandlers();  // Main-thread-only.
@@ -174,20 +171,7 @@ function setupDebugEnv() {
   }
 }
 
-function setupSignalHandlers() {
-  const {
-    createSignalHandlers
-  } = require('internal/process/main_thread_only');
-  const {
-    startListeningIfSignal,
-    stopListeningIfSignal
-  } = createSignalHandlers();
-  process.on('newListener', startListeningIfSignal);
-  process.on('removeListener', stopListeningIfSignal);
-}
-
-// This has to be called after both initializeReport() and
-// setupSignalHandlers() are called
+// This has to be called after initializeReport() is called
 function initializeReportSignalHandlers() {
   if (!getOptionValue('--experimental-report')) {
     return;
